requirements.txt
python-dotenv
python-telegram-bot
emoji
dice
cochar